+++ lib/data/models/product_model.dart
@@ -1,73 +1,3 @@
-// import 'package:course/data/models/category_model.dart';
-
-// class ProductModel {
-//   CategoryModel? category;
-//   int? id;
-//   String? name;
-//   String? nameAr;
-//   String? description;
-//   String? descriptionAr;
-//   String? image;
-//   int? quantity;
-//   int? isActive;
-//   int? price;
-//   int? discount;
-//   String? createdAt;
-//   String? updatedAt;
-
-//   ProductModel(
-//       {this.category,
-//       this.name,
-//       this.nameAr,
-//       this.description,
-//       this.descriptionAr,
-//       this.image,
-//       this.quantity,
-//       this.isActive,
-//       this.price,
-//       this.discount,
-//       this.createdAt,
-//       this.updatedAt});
-
-//   ProductModel.fromJson(Map<String, dynamic> json) {
-//     category = json['category'] != null
-//         ? CategoryModel.fromJson(json['category'])
-//         : null;
-//     id = json['id'];
-//     name = json['name'];
-//     nameAr = json['name_ar'];
-//     description = json['description'];
-//     descriptionAr = json['description_ar'];
-//     image = json['image'];
-//     quantity = json['quantity'];
-//     isActive = json['is_active'];
-//     price = json['price'];
-//     discount = json['discount'];
-//     createdAt = json['created_at'];
-//     updatedAt = json['updated_at'];
-//   }
-
-//   Map<String, dynamic> toJson() {
-//     final Map<String, dynamic> data = new Map<String, dynamic>();
-//     if (this.category != null) {
-//       data['category'] = this.category!.toJson();
-//     }
-//     data['id'] = this.id;
-//     data['name'] = this.name;
-//     data['name_ar'] = this.nameAr;
-//     data['description'] = this.description;
-//     data['description_ar'] = this.descriptionAr;
-//     data['image'] = this.image;
-//     data['quantity'] = this.quantity;
-//     data['is_active'] = this.isActive;
-//     data['price'] = this.price;
-//     data['discount'] = this.discount;
-//     data['created_at'] = this.createdAt;
-//     data['updated_at'] = this.updatedAt;
-//     return data;
-//   }
-// }
-
 class ProductModel {
   int? id;
   String? name;
@@ -77,7 +7,7 @@ class ProductModel {
   String? image;
   int? quantity;
   int? isActive;
-  int? price;
+  double? price;
   int? discount;
   int? categoryId;
   String? categoryName;
@@ -111,7 +41,7 @@ class ProductModel {
     image = json['image'];
     quantity = json['quantity'];
     isActive = json['is_active'];
-    price = json['price'];
+    price = json['price'].toDouble();
     discount = json['discount'];
     categoryId = json['category_id'];
     categoryName = json['category_name'];
